refactor: remove reflection when storing CBOR on decode (#983)
This removes multiple uses of reflection during decoding by creating a
static type of the same shape in UnmarshalCBOR. This works because the
UnmarshalCBOR function from the original type does not follow to the new
type, so generic decoding behavior is used instead.

This was only possible after removing embedding in any of the types that
we want to capture the original CBOR on. The promoted UnmarshalCBOR
function from the embedded type(s) were being called otherwise, which
did not produce the desired result.

Signed-off-by: Aurora Gaffney <aurora@blinklabs.io>