feat(ledger): Add RawScriptBytes to Script interface (#1210)
It's rather awkward to get the script bytes right now, as you have to try casting it to different script types; If you just need the bytes (ex: to store those in a database, etc.) then the RawScriptBytes method gives you access directly to those bytes (without the version wrapper). Might also consider adding a RawBytes to give *exactly* what's on chain, and/or a Version, to indicate which script version it is, but I held off on that for now as this is all that I need to unblock myself.

Signed-off-by: Pi Lanningham <pi@sundae.fi>